test: stabilise playwright setup for PR 5217
